fix node test
