+++ ext/node/lib.rs
@@ -674,8 +674,6 @@ deno_core::extension!(deno_node,
     "path/mod.ts",
     "path/separator.ts",
     "readline/promises.ts",
-    "timers.ts",
-    "timers/promises.ts",
     "node:_http_agent" = "_http_agent.mjs",
     "node:_http_common" = "_http_common.ts",
     "node:_http_outgoing" = "_http_outgoing.ts",
